% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circlesplot.R
\name{circlesplot}
\alias{circlesplot}
\title{Plots a circlesplot based upon the given input}
\usage{
circlesplot(
  cp_vals = NULL,
  cp_text = NULL,
  cp_max = 10L,
  cp_line_width = 2L,
  cp_title = "",
  cp_color = NULL,
  cp_title_size = 1.5
)
}
\arguments{
\item{cp_title_size}{}
}
\value{
Nothing
}
\description{
Plots a circlesplot based upon the given input
}
\examples{

colors = c('#D1BBD7', '#AE76A3', '#882E72', '#1965B0', '#5289C7', '#7BAFDE', '#4EB265', '#90C987')
circlesplot(cp_vals=c(5,5,4,5,5,5,2,1), cp_text=c('8','7','6','5','4','3','2','1'), cp_max=3L, cp_title="Some title", cp_color=colors)


# For coloring, you can also use viridis package:
library("viridis")
circlesplot(cp_vals=c(5,5,4,5,5,5,2,1), cp_text=c('8','7','6','5','4','3','2','1'), cp_max=3L, cp_title="Some title", cp_color=viridis(8))

}
